+++ src/c11threads.c
@@ -56,7 +56,7 @@ int thrd_join(thrd_t thr, int *res) {
         return thrd_error;
     }
     if (res) {
-        *res = (long)retval;
+        *res = (int) retval;
     }
     return thrd_success;
 }
